Flush level (part 2)
This phase will incorporate a new Flush Level to the game. In this
level, the grid should be the same size as the Same Rank Trio level (5
rows by 10 columns). However, players uncover five cards instead of
three on each turn. A wining hand consists of all five cards having the
same suit. The score for each hand should be computed as 700 points
plus the sum of the ranks in all the cards. For instance, a hand with
2, 5, 10, K, and A of clubs will accumulate 700 + 2 + 5 + 12 + 13 for a
total of 732. A 5-point penalty should be deducted from the score for
each unsuccessful turn taken. Remember to update the Help page to
include a clear and concise description of the level.